Исправлено: проверка ссылок в заголовке сайта; Добавлено: взятие скриншота текущей карты; В планах: брать скриншоты всех карт в соответствии с изменением времени.
