+++ src/main/java/webpages/gismeteo/pages/NewsPageGismeteo.java
@@ -9,7 +9,7 @@ public class NewsPageGismeteo extends DefaultPage {
         super(driver);
     }
 
-    public static final By linksList = new By.ByCssSelector("body > header > div.header-subnav > div.header-container.wrap");
+    public static final By linksList = new By.ByCssSelector("body > header > div.header-subnav > div.header-container.wrap > div > div");
     public static final By readAlsoElementSelector = new By.ByCssSelector("body > section > div.content-column.column1 > section.section.section-content.section-visible > div.articles-items.js-articles-items");
     public static final By explainingRFTabSelector = new By.ByCssSelector("body > section > div.content-column.column2 > div > section:nth-child(1) > div.feed.feed-partner");
     public static final By twentyFourSMITabSelector = new By.ByCssSelector("body > section > div.content-column.column2 > div > section:nth-child(2) > div.feed.feed-partner");
