+++ src/test/java/IndexTest/ParticularTests/NewsTests/AnimalsNewsPageTest.java
@@ -1,4 +1,4 @@
-package IndexTest.ParticularTests;
+package IndexTest.ParticularTests.NewsTests;
 
 import IndexTest.DefaultPageTest;
 import common.ConfiguresAndConstants;
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class WildfireNewsPageTest extends DefaultPageTest {
+public class AnimalsNewsPageTest extends DefaultPageTest {
 /*
     @BeforeEach
     public void beforeEachMethod() {
@@ -25,15 +25,15 @@ public class WildfireNewsPageTest extends DefaultPageTest {
     }*/
 
     @Test
-    public void checkWildfireURL() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire");
-        Assertions.assertEquals("https://www.gismeteo.ru/news/wildfire/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/wildfire/" + driver.getCurrentUrl() + " не соответствуют");
+    public void checkAnimalsURL() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+        Assertions.assertEquals("https://www.gismeteo.ru/news/animals/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/animals/" + driver.getCurrentUrl() + " не соответствуют");
         driver.quit();
     }
 
     @Test
-    public void wildfireCheckTopLinks() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void animalsCheckTopLinks() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         ArrayList<String> topLinksSet = new ArrayList<>();
         topLinksSet.add("Все новости");
         topLinksSet.add("Авто");
@@ -49,8 +49,8 @@ public class WildfireNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void wildfireNewsCountElementsFirstColumn() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void animalsNewsCountElementsFirstColumn() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         WebElement firstColumnTab = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         String amountOfChildren = "10";
         Assertions.assertEquals(amountOfChildren, firstColumnTab.getAttribute("childElementCount"), "Количество детей (" + firstColumnTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -58,20 +58,20 @@ public class WildfireNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesTopSevenWildfireNewsTab() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void getTitlesTopSevenAnimalsNewsTab() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.topSevenNewsSelector);
         List<WebElement> someElements = someElement.findElements(By.className("card-wrap"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void readAlsoCounterChildrenWildfireNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void readAlsoCounterChildrenAnimalsNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         String amountOfChildren = "10";
         WebElement readAlsoField = driver.findElement(NewsPageGismeteo.readAlsoElementSelector);
         Assertions.assertEquals(amountOfChildren, readAlsoField.getAttribute("childElementCount"), "Количество детей (" + readAlsoField.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -79,8 +79,8 @@ public class WildfireNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void explainingRFCountChildrenWildfireNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void explainingRFCountChildrenAnimalsNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         String amountOfChildren = "9";
         WebElement explainingRFTab = driver.findElement(NewsPageGismeteo.explainingRFTabSelector);
         Assertions.assertEquals(amountOfChildren, explainingRFTab.getAttribute("childElementCount"), "Количество детей (" + explainingRFTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -88,8 +88,8 @@ public class WildfireNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void TwentyFourSMICountChildrenWildfireNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void TwentyFourSMICountChildrenAnimalsNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         String amountOfChildren = "11";
         WebElement twentySMIFourTab = driver.findElement(NewsPageGismeteo.twentyFourSMITabSelector);
         Assertions.assertEquals(amountOfChildren, twentySMIFourTab.getAttribute("childElementCount"), "Количество детей (" + twentySMIFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -97,8 +97,8 @@ public class WildfireNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void MoscowTwentyFourCountChildrenWildfireNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void MoscowTwentyFourCountChildrenAnimalsNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         String amountOfChildren = "8";
         WebElement moscowTwentyFourFourTab = driver.findElement(NewsPageGismeteo.moscowTwentyFourTabSelector);
         Assertions.assertEquals(amountOfChildren, moscowTwentyFourFourTab.getAttribute("childElementCount"), "Количество детей (" + moscowTwentyFourFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -106,35 +106,35 @@ public class WildfireNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesReadAlsoTabWildfireNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void getTitlesReadAlsoTabAnimalsNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void checkMoreNewsButtonWildfireNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+    public void checkMoreNewsButtonAnimalsNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void clickMoreNewsButtonWildfireNews() {
+    public void clickMoreNewsButtonAnimalsNews() {
         int maxAmountOfClicks = 100;
         int amountOfClicks = 0;
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
         WebElement moreNewsButton = driver.findElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         while (moreNewsButton.isDisplayed() && amountOfClicks <= maxAmountOfClicks) {
             clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
@@ -143,7 +143,7 @@ public class WildfireNewsPageTest extends DefaultPageTest {
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
