+++ src/test/java/IndexTest/ParticularTests/NewsTests/AutoNewsPageTest.java
@@ -1,4 +1,4 @@
-package IndexTest.ParticularTests;
+package IndexTest.ParticularTests.NewsTests;
 
 import IndexTest.DefaultPageTest;
 import common.ConfiguresAndConstants;
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class AnimalsNewsPageTest extends DefaultPageTest {
+public class AutoNewsPageTest extends DefaultPageTest {
 /*
     @BeforeEach
     public void beforeEachMethod() {
@@ -25,15 +25,15 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
     }*/
 
     @Test
-    public void checkAnimalsURL() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
-        Assertions.assertEquals("https://www.gismeteo.ru/news/animals/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/animals/" + driver.getCurrentUrl() + " не соответствуют");
+    public void checkAutoURL() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto");
+        Assertions.assertEquals("https://www.gismeteo.ru/news/auto/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/auto/" + driver.getCurrentUrl() + " не соответствуют");
         driver.quit();
     }
 
     @Test
-    public void animalsCheckTopLinks() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void checkTopLinks() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         ArrayList<String> topLinksSet = new ArrayList<>();
         topLinksSet.add("Все новости");
         topLinksSet.add("Авто");
@@ -49,8 +49,8 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void animalsNewsCountElementsFirstColumn() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void autoNewsCountElementsFirstColumn() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         WebElement firstColumnTab = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         String amountOfChildren = "10";
         Assertions.assertEquals(amountOfChildren, firstColumnTab.getAttribute("childElementCount"), "Количество детей (" + firstColumnTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -58,20 +58,20 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesTopSevenAnimalsNewsTab() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void getTitlesTopSevenAutoNewsTab() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.topSevenNewsSelector);
         List<WebElement> someElements = someElement.findElements(By.className("card-wrap"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void readAlsoCounterChildrenAnimalsNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void readAlsoCounterChildrenAutoNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         String amountOfChildren = "10";
         WebElement readAlsoField = driver.findElement(NewsPageGismeteo.readAlsoElementSelector);
         Assertions.assertEquals(amountOfChildren, readAlsoField.getAttribute("childElementCount"), "Количество детей (" + readAlsoField.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -79,8 +79,8 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void explainingRFCountChildrenAnimalsNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void explainingRFCountChildrenAutoNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         String amountOfChildren = "9";
         WebElement explainingRFTab = driver.findElement(NewsPageGismeteo.explainingRFTabSelector);
         Assertions.assertEquals(amountOfChildren, explainingRFTab.getAttribute("childElementCount"), "Количество детей (" + explainingRFTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -88,8 +88,8 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void TwentyFourSMICountChildrenAnimalsNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void TwentyFourSMICountChildrenAutoNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         String amountOfChildren = "11";
         WebElement twentySMIFourTab = driver.findElement(NewsPageGismeteo.twentyFourSMITabSelector);
         Assertions.assertEquals(amountOfChildren, twentySMIFourTab.getAttribute("childElementCount"), "Количество детей (" + twentySMIFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -97,8 +97,8 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void MoscowTwentyFourCountChildrenAnimalsNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void MoscowTwentyFourCountChildrenAutoNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         String amountOfChildren = "8";
         WebElement moscowTwentyFourFourTab = driver.findElement(NewsPageGismeteo.moscowTwentyFourTabSelector);
         Assertions.assertEquals(amountOfChildren, moscowTwentyFourFourTab.getAttribute("childElementCount"), "Количество детей (" + moscowTwentyFourFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -106,35 +106,35 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesReadAlsoTabAnimalsNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void getTitlesReadAlsoTabAutoNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void checkMoreNewsButtonAnimalsNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+    public void checkMoreNewsButtonAutoNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void clickMoreNewsButtonAnimalsNews() {
+    public void clickMoreNewsButtonAutoNews() {
         int maxAmountOfClicks = 100;
         int amountOfClicks = 0;
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/animals/");
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
         WebElement moreNewsButton = driver.findElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         while (moreNewsButton.isDisplayed() && amountOfClicks <= maxAmountOfClicks) {
             clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
@@ -143,7 +143,7 @@ public class AnimalsNewsPageTest extends DefaultPageTest {
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/animals/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
