+++ src/test/java/IndexTest/ParticularTests/NewsTests/CoronavirusNewsPageTest.java
@@ -1,4 +1,4 @@
-package IndexTest.ParticularTests;
+package IndexTest.ParticularTests.NewsTests;
 
 import IndexTest.DefaultPageTest;
 import common.ConfiguresAndConstants;
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class AutoNewsPageTest extends DefaultPageTest {
+public class CoronavirusNewsPageTest extends DefaultPageTest {
 /*
     @BeforeEach
     public void beforeEachMethod() {
@@ -25,15 +25,15 @@ public class AutoNewsPageTest extends DefaultPageTest {
     }*/
 
     @Test
-    public void checkAutoURL() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto");
-        Assertions.assertEquals("https://www.gismeteo.ru/news/auto/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/auto/" + driver.getCurrentUrl() + " не соответствуют");
+    public void checkCoronavirusURL() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus");
+        Assertions.assertEquals("https://www.gismeteo.ru/news/coronavirus/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/coronavirus/" + driver.getCurrentUrl() + " не соответствуют");
         driver.quit();
     }
 
     @Test
-    public void checkTopLinks() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void coronavirusCheckTopLinks() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         ArrayList<String> topLinksSet = new ArrayList<>();
         topLinksSet.add("Все новости");
         topLinksSet.add("Авто");
@@ -49,8 +49,8 @@ public class AutoNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void autoNewsCountElementsFirstColumn() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void coronavirusNewsCountElementsFirstColumn() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         WebElement firstColumnTab = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         String amountOfChildren = "10";
         Assertions.assertEquals(amountOfChildren, firstColumnTab.getAttribute("childElementCount"), "Количество детей (" + firstColumnTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -58,20 +58,20 @@ public class AutoNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesTopSevenAutoNewsTab() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void getTitlesTopSevenCoronavirusNewsTab() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.topSevenNewsSelector);
         List<WebElement> someElements = someElement.findElements(By.className("card-wrap"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/coronavirus/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void readAlsoCounterChildrenAutoNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void readAlsoCounterChildrenCoronavirusNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         String amountOfChildren = "10";
         WebElement readAlsoField = driver.findElement(NewsPageGismeteo.readAlsoElementSelector);
         Assertions.assertEquals(amountOfChildren, readAlsoField.getAttribute("childElementCount"), "Количество детей (" + readAlsoField.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -79,8 +79,8 @@ public class AutoNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void explainingRFCountChildrenAutoNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void explainingRFCountChildrenCoronavirusNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         String amountOfChildren = "9";
         WebElement explainingRFTab = driver.findElement(NewsPageGismeteo.explainingRFTabSelector);
         Assertions.assertEquals(amountOfChildren, explainingRFTab.getAttribute("childElementCount"), "Количество детей (" + explainingRFTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -88,8 +88,8 @@ public class AutoNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void TwentyFourSMICountChildrenAutoNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void TwentyFourSMICountChildrenCoronavirusNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         String amountOfChildren = "11";
         WebElement twentySMIFourTab = driver.findElement(NewsPageGismeteo.twentyFourSMITabSelector);
         Assertions.assertEquals(amountOfChildren, twentySMIFourTab.getAttribute("childElementCount"), "Количество детей (" + twentySMIFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -97,8 +97,8 @@ public class AutoNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void MoscowTwentyFourCountChildrenAutoNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void MoscowTwentyFourCountChildrenCoronavirusNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         String amountOfChildren = "8";
         WebElement moscowTwentyFourFourTab = driver.findElement(NewsPageGismeteo.moscowTwentyFourTabSelector);
         Assertions.assertEquals(amountOfChildren, moscowTwentyFourFourTab.getAttribute("childElementCount"), "Количество детей (" + moscowTwentyFourFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -106,35 +106,35 @@ public class AutoNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesReadAlsoTabAutoNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void getTitlesReadAlsoTabCoronavirusNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/coronavirus/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void checkMoreNewsButtonAutoNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+    public void checkMoreNewsButtonCoronavirusNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/coronavirus/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void clickMoreNewsButtonAutoNews() {
+    public void clickMoreNewsButtonCoronavirusNews() {
         int maxAmountOfClicks = 100;
         int amountOfClicks = 0;
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/auto/");
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/coronavirus/");
         WebElement moreNewsButton = driver.findElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         while (moreNewsButton.isDisplayed() && amountOfClicks <= maxAmountOfClicks) {
             clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
@@ -143,7 +143,7 @@ public class AutoNewsPageTest extends DefaultPageTest {
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/auto/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/coronavirus/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
