+++ src/test/java/IndexTest/ParticularTests/NewsTests/WeatherNewsPageTest.java
@@ -1,4 +1,4 @@
-package IndexTest.ParticularTests;
+package IndexTest.ParticularTests.NewsTests;
 
 import IndexTest.DefaultPageTest;
 import common.ConfiguresAndConstants;
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class ScienceNewsPageTest extends DefaultPageTest {
+public class WeatherNewsPageTest extends DefaultPageTest {
 /*
     @BeforeEach
     public void beforeEachMethod() {
@@ -25,15 +25,15 @@ public class ScienceNewsPageTest extends DefaultPageTest {
     }*/
 
     @Test
-    public void checkScienceURL() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
-        Assertions.assertEquals("https://www.gismeteo.ru/news/science/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/science/" + driver.getCurrentUrl() + " не соответствуют");
+    public void checkWeatherURL() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather");
+        Assertions.assertEquals("https://www.gismeteo.ru/news/weather/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/weather/" + driver.getCurrentUrl() + " не соответствуют");
         driver.quit();
     }
 
     @Test
-    public void natureCheckTopLinks() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void weatherCheckTopLinks() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         ArrayList<String> topLinksSet = new ArrayList<>();
         topLinksSet.add("Все новости");
         topLinksSet.add("Авто");
@@ -49,8 +49,8 @@ public class ScienceNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void scienceNewsCountElementsFirstColumn() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void weatherNewsCountElementsFirstColumn() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         WebElement firstColumnTab = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         String amountOfChildren = "10";
         Assertions.assertEquals(amountOfChildren, firstColumnTab.getAttribute("childElementCount"), "Количество детей (" + firstColumnTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -58,20 +58,20 @@ public class ScienceNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesTopSevenScienceNewsTab() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void getTitlesTopSevenWeatherNewsTab() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.topSevenNewsSelector);
         List<WebElement> someElements = someElement.findElements(By.className("card-wrap"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/science/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/weather/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void readAlsoCounterChildrenScienceNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void readAlsoCounterChildrenWeatherNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         String amountOfChildren = "10";
         WebElement readAlsoField = driver.findElement(NewsPageGismeteo.readAlsoElementSelector);
         Assertions.assertEquals(amountOfChildren, readAlsoField.getAttribute("childElementCount"), "Количество детей (" + readAlsoField.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -79,8 +79,8 @@ public class ScienceNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void explainingRFCountChildrenScienceNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void explainingRFCountChildrenWeatherNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         String amountOfChildren = "9";
         WebElement explainingRFTab = driver.findElement(NewsPageGismeteo.explainingRFTabSelector);
         Assertions.assertEquals(amountOfChildren, explainingRFTab.getAttribute("childElementCount"), "Количество детей (" + explainingRFTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -88,8 +88,8 @@ public class ScienceNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void TwentyFourSMICountChildrenScienceNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void TwentyFourSMICountChildrenWeatherNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         String amountOfChildren = "11";
         WebElement twentySMIFourTab = driver.findElement(NewsPageGismeteo.twentyFourSMITabSelector);
         Assertions.assertEquals(amountOfChildren, twentySMIFourTab.getAttribute("childElementCount"), "Количество детей (" + twentySMIFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -97,8 +97,8 @@ public class ScienceNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void MoscowTwentyFourCountChildrenScienceNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void MoscowTwentyFourCountChildrenWeatherNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         String amountOfChildren = "8";
         WebElement moscowTwentyFourFourTab = driver.findElement(NewsPageGismeteo.moscowTwentyFourTabSelector);
         Assertions.assertEquals(amountOfChildren, moscowTwentyFourFourTab.getAttribute("childElementCount"), "Количество детей (" + moscowTwentyFourFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -106,35 +106,35 @@ public class ScienceNewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesReadAlsoTabScienceNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void getTitlesReadAlsoTabWeatherNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/science/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/weather/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void checkMoreNewsButtonScienceNews() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+    public void checkMoreNewsButtonWeatherNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/science/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/weather/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void clickMoreNewsButtonScienceNews() {
+    public void clickMoreNewsButtonWeatherNews() {
         int maxAmountOfClicks = 100;
         int amountOfClicks = 0;
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/science/");
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/weather/");
         WebElement moreNewsButton = driver.findElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         while (moreNewsButton.isDisplayed() && amountOfClicks <= maxAmountOfClicks) {
             clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
@@ -143,7 +143,7 @@ public class ScienceNewsPageTest extends DefaultPageTest {
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/science/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/weather/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
