+++ src/test/java/IndexTest/ParticularTests/NewsTests/WildfireNewsPageTest.java
@@ -1,4 +1,4 @@
-package IndexTest.ParticularTests;
+package IndexTest.ParticularTests.NewsTests;
 
 import IndexTest.DefaultPageTest;
 import common.ConfiguresAndConstants;
@@ -8,9 +8,11 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import webpages.gismeteo.pages.NewsPageGismeteo;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
-public class NewsPageTest extends DefaultPageTest {
+public class WildfireNewsPageTest extends DefaultPageTest {
 /*
     @BeforeEach
     public void beforeEachMethod() {
@@ -23,16 +25,16 @@ public class NewsPageTest extends DefaultPageTest {
     }*/
 
     @Test
-    public void checkNewsURL() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
-        Assertions.assertEquals("https://www.gismeteo.ru/news/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/" + driver.getCurrentUrl() + " не соответствуют");
+    public void checkWildfireURL() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire");
+        Assertions.assertEquals("https://www.gismeteo.ru/news/wildfire/", driver.getCurrentUrl(), "https://www.gismeteo.ru/news/wildfire/" + driver.getCurrentUrl() + " не соответствуют");
         driver.quit();
     }
 
     @Test
-    public void checkTopLinks() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
-        ArrayList<String> topLinksSet = new ArrayList<String>();
+    public void wildfireCheckTopLinks() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+        ArrayList<String> topLinksSet = new ArrayList<>();
         topLinksSet.add("Все новости");
         topLinksSet.add("Авто");
         topLinksSet.add("Животные");
@@ -47,21 +49,8 @@ public class NewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesTopSevenNewsTab() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
-        WebElement someElement = driver.findElement(NewsPageGismeteo.topSevenNewsSelector);
-        List<WebElement> someElements = someElement.findElements(By.className("card-wrap"));
-        for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/", articleItem.getAttribute("baseURI"), "URLs do not match");
-            Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
-        }
-        driver.quit();
-    }
-
-
-    @Test
-    public void countElementsFirstColumn() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void wildfireNewsCountElementsFirstColumn() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         WebElement firstColumnTab = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         String amountOfChildren = "10";
         Assertions.assertEquals(amountOfChildren, firstColumnTab.getAttribute("childElementCount"), "Количество детей (" + firstColumnTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -69,8 +58,20 @@ public class NewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void readAlsoCounterChildren() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void getTitlesTopSevenWildfireNewsTab() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
+        WebElement someElement = driver.findElement(NewsPageGismeteo.topSevenNewsSelector);
+        List<WebElement> someElements = someElement.findElements(By.className("card-wrap"));
+        for (WebElement articleItem : someElements) {
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
+        }
+        driver.quit();
+    }
+
+    @Test
+    public void readAlsoCounterChildrenWildfireNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         String amountOfChildren = "10";
         WebElement readAlsoField = driver.findElement(NewsPageGismeteo.readAlsoElementSelector);
         Assertions.assertEquals(amountOfChildren, readAlsoField.getAttribute("childElementCount"), "Количество детей (" + readAlsoField.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -78,8 +79,8 @@ public class NewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void explainingRFCountChildren() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void explainingRFCountChildrenWildfireNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         String amountOfChildren = "9";
         WebElement explainingRFTab = driver.findElement(NewsPageGismeteo.explainingRFTabSelector);
         Assertions.assertEquals(amountOfChildren, explainingRFTab.getAttribute("childElementCount"), "Количество детей (" + explainingRFTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -87,8 +88,8 @@ public class NewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void TwentyFourSMICountChildren() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void TwentyFourSMICountChildrenWildfireNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         String amountOfChildren = "11";
         WebElement twentySMIFourTab = driver.findElement(NewsPageGismeteo.twentyFourSMITabSelector);
         Assertions.assertEquals(amountOfChildren, twentySMIFourTab.getAttribute("childElementCount"), "Количество детей (" + twentySMIFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -96,8 +97,8 @@ public class NewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void MoscowTwentyFourCountChildren() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void MoscowTwentyFourCountChildrenWildfireNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         String amountOfChildren = "8";
         WebElement moscowTwentyFourFourTab = driver.findElement(NewsPageGismeteo.moscowTwentyFourTabSelector);
         Assertions.assertEquals(amountOfChildren, moscowTwentyFourFourTab.getAttribute("childElementCount"), "Количество детей (" + moscowTwentyFourFourTab.getAttribute("childElementCount") + ") не равно ожидаемому (" + amountOfChildren + ")");
@@ -105,40 +106,35 @@ public class NewsPageTest extends DefaultPageTest {
     }
 
     @Test
-    public void getTitlesReadAlsoTab() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void getTitlesReadAlsoTabWildfireNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
         }
         driver.quit();
     }
 
     @Test
-    public void checkMoreNewsButton() {
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void checkMoreNewsButtonWildfireNews() {
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
-        Integer countArticles = 0;
         for (WebElement articleItem : someElements) {
-            System.out.println(articleItem.getAttribute("innerText"));
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
-            countArticles++;
         }
-        System.out.println(countArticles);
         driver.quit();
     }
 
     @Test
-    public void clickMoreNewsButton() {
-        Integer maxAmountOfClicks = 100;
-        Integer amountOfClicks = 0;
-        Integer countArticles = 0;
-        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/");
+    public void clickMoreNewsButtonWildfireNews() {
+        int maxAmountOfClicks = 100;
+        int amountOfClicks = 0;
+        defaultPage.openWebPages(ConfiguresAndConstants.defaultURL + "news/wildfire/");
         WebElement moreNewsButton = driver.findElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
         while (moreNewsButton.isDisplayed() && amountOfClicks <= maxAmountOfClicks) {
             clickElement(NewsPageGismeteo.readMoreNewsFirstColumnTabSelector);
@@ -147,9 +143,8 @@ public class NewsPageTest extends DefaultPageTest {
         WebElement someElement = driver.findElement(NewsPageGismeteo.firstColumnTabSelector);
         List<WebElement> someElements = someElement.findElements(By.className("article-item"));
         for (WebElement articleItem : someElements) {
-            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/", articleItem.getAttribute("baseURI"), "URLs do not match");
+            Assertions.assertEquals(ConfiguresAndConstants.defaultURL + "news/wildfire/", articleItem.getAttribute("baseURI"), "URLs do not match");
             Assertions.assertEquals(articleItem.getAttribute("innerText"), articleItem.getAttribute("outerText"));
-            countArticles++;
         }
         driver.quit();
     }
